feat(router): Added automatic response type resolution
The default response type is now determined by the return type of the routes callable.  This will of course default to HTML, but if there is a match between the response class and a response type (such as Hazaar\Controller\Response\JSON and TEXT_JSON) then that type will be used as the response type for errors and exceptions.