feat(dbi): Added DBI checkpoints
Checkpoints will consolidate all migration files into a single file.  It does this by replaying the migrations into a schema (the same way an initial schema is applied) and the converting that back into a migration that is written to one file.  All migrations are then removed and the schema_version table reset.