fix(smarty): merge functions from includes
We had a problem where if a file was included within another include and that file contained functions, those functions were not accessible to the including file.  I have fixed this by merging functions from included files into the calling object so that each time a file is included, the functions are merged all the way back up to the root object.