fix(smarty): random spaces during render
When rendering on the CLI the process would output random spaces.  I'm still not sure why as everything should be inside an output buffer, but it was something to do with the default value.  So I removed that feature as it's not really used anymore and now the write() function will only echo if the value is not null.