fix(dbi): changed querybuilder init.
The SQL trait now maintains the queryBuilder.  It now needs to be initialised by the driver before it can be used.

It is now a singleton, but this might change if it ends up causing issues.  The reason it is a singleton now is because the schema name needs to be set up when the driver initialises.  If this is a problem, then we might store the schema name and create a new instance when needed.