chore(router): renamed controller run method
Renamed Controller::run() to Controller::runRoute().  This method runs a route but I wanted to use Controller::run() as a direct exec function available for custom controllers.

Updated everything to use the new runRoute() method except Diagnostic controllers which use run() because that is what it is designed for, being passed directly to the application to just run it.