fix(dev): changed to PHPTools extension and fixed issues
Switched to DEVSENSE.phptools-vscode for handling intelesense and running unit tests.  This extension flagged an issue with code standards where sometimes I use relative references in interface implements.  I usually do this when the interface is the same name as the class it is being used with.  This fixes these issues.