feat(smarty): enhanced renderer
The renderer is now in it's own class instead of being stupidly defined as a string and processed with eval().  We still need to do this, but now it is the absolute minimum that is just a render() function that is defined in a child class as a string.  This means that all the work is done by the new Hazaar\Template\Smarty\Renderer class which makes things sooooo much easier to debug.