chore(view): refactored the view function handler layout/url methods
The view function handler defines the functions that can be called while rendering a view.  These are normally implemented directly on the class itself, but for the methods that are normally handled by the view/layout classes, this object class has become a gatekeeper so that it is no longer possible to try and call functions that shouldn't be called.