feat(smarty): enhanced includes
Includes are now compiled and rendered using an entirely new Smarty object.  The reason for this is that it allows the file to be rendered EXACTLY as it would be if called directly.  This includes creating any cache files, checking for modifications, etc.

So this solves the problem with include file changes not being detected and triggering a recompile.  It also means that we can include the same file multiple times as needed which was also previously a problem.