fix(smarty): removed renderer type
Rendering API docs has caused me to get rid of this as the criteria is a bit more limited and definable.  Basically, if the source content is a file and we have a runtime dir, then we should use the PHP renderer.  Otherwise use eval().  I have removed the pre-defined type and also the exception because as a fallback we will just always use the eval() method.