fix(smarty): performance tweaks
I've removed a few things from the predefined params which can have a performance impact, such as access to env(), constants and the application config which was being converted to an array for every render.

Removing all this caused a 5ms page render with a few includes to drop to <2ms.  Even a straight dump() call can render in about <0.5ms.  Pretty happy with that.