chore(ake): DEPRECATED ake()
ake() was great back when PHP was shit.  but PHP is actually really nice these days and ake() is not needed thank to the null coalescing operator.

ake() has been marked as deprecated but still exists for now.  Despite that, I have removed all internal uses of ake() and replaced with null coalescing.