chore(smarty): Include file indexing
Previously when a smarty file was included it was instantiating a new Smarty object, setting it up and rendering it.  This was still pretty quick because if the file was cached, it didn't need to compile it again.

However, with in-memory renders, the compile/render process uses eval() which is quite slow.  So now, we only create a Smarty object once for each file and we store it on an index.  It is safe to use it again with new data without the need to re-compile and re-eval() anything.

Performance on API doc gen has gone from 8-10 seconds to render docs down to 2 seconds.  Faaaark.