feat(cli): inserted modules concept to allow for multiple commands
A 'command' is now a module and these modules can register multiple commands.

I'm using this in the DocModule because I want it to handle both the 'doc' command and the 'index' command.  This will allow the prepare to run some common code and then each method can be different.