fix(ci): OOMKilled issue
This was happening because whenever we throw an error and insane amount of output was trying to be written to the console because xdebug safety constraints had been turned off.  I have turned them back to the defaults so now we should just get a nice little error and that's it.

I have also added memory usage log output to the Documentor.