feat(cache): file cleanup on shutdown based on namespace timeout
This is a better implementation.  Now we can look at the namespace timeout value in the database file and if has expired, plus a 5 minute window, we remove the file.  The 5 minute window should be more than enough to prevent collisions.

The flush option still sets how long until a file is checked to prevent unnecessarily opening DB files.