feat(dbi): querybuilder prepare updates
This commit has the changes needed to get the query builder working with prepared statements for the SELECT only.  This primarily is getting the criteria actions working.

There is now a value index to handle cases where the same field is referenced multiple times with the same or different values.  if the value is the same then the placeholder tag is re-used.  If they are different, then the tag is changed by incrementing a counter suffix.