feat(util): Adds GeoData module to cache database file
Adds a new GeoData module that enables caching of the GeoData database file.

This feature allows users to cache the geodata database file locally,
improving performance and reducing reliance on external data sources.
The module includes a command to fetch and cache the database file,
as well as display relevant information such as file size and location.

Also refactors the GeoData class to be more flexible and to fetch the DB file.