feat(cli): Adds config management module
Adds a new console module for managing application configurations.

This module allows users to view, check, get, and set configuration
values via the command line. It also includes functionality to
locate the application path and load the environment configuration.

Also adds `envsubst` to the Dockerfile and removes it from the CI,
and renames the "include" key in the application config to "import".