fix(util): fixed btree balance issue losing nodes.
This was happening when a node was being split during a set.  The set node was chosen first, then the split was performed, and if the set key should have ended up on the new node, it was incorrectly written to the already selected node.

The fix was to do the split right after adding a new key.